Improve how _get_temperature_index treats boundary
When EigenData2._get_temperature_index had an input that equals the
boundary, it had been issuing a warning.  I changed this so that the
warning is not issued when the temperature equals the minimum or
maximum temperature in the grid.  When I start a simulation at 10000
K, I no longer get a warning.