Extrapolate n and T outside of time interval
If the temperature or number density is inputted as an array, then a
problem may arise if the time is not within time_start or time_max
(e.g., as a result of roundoff error).  I set interp1d to extrapolate
outside of the time interval.  I changed in_time_interval to allow for
a time buffer (e.g., to allow the time to be slightly outside of the
time interval to account for roundoff error).  Warnings should be
raised if the time is outside of the time interval.  I also used
astropy.units.quantity_input to decorate some of the methods.